+++ src/main/java/kusitms/backend/auth/jwt/JWTFilter.java
@@ -51,6 +51,7 @@ public class JWTFilter extends OncePerRequestFilter {
                 || path.equals("/api/v1/token/re-issue")
                 || path.equals("/api/v1/test/docs") || path.startsWith(("/swagger-ui")) || path.startsWith("/v3/api-docs")
                 || path.startsWith("/api/v1/results") || path.startsWith("/api/v1/stadium")
+                || path.startsWith("/api/v1/culture")
                 || path.startsWith("/api/v1/chatbot")
                 ;
     }
