+++ src/main/java/kusitms/backend/culture/status/FoodSuccessStatus.java
@@ -1,4 +1,4 @@
-package kusitms.backend.food.status;
+package kusitms.backend.culture.status;
 
 import kusitms.backend.global.code.BaseCode;
 import kusitms.backend.global.dto.ReasonDto;
